+++ CHANGELOG.md
@@ -11,4 +11,11 @@ Possible categories for the changes: Updated, Fixed, Features, Deprecated
 ###Android
 
 ####	 Updated
+
 * Base interactor not handling exceptions outside its domain anymore
+
+###iOS
+
+#### Updated
+* Expose user object in Session context instead of raw attributtes
+